add restore  password
